incorporate make_schema into mysql module/object
